backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
websocket-client==1.8.0
requests==2.32.3
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
pyaxmlparser==0.3.31
python-multipart==0.0.20
telethon==1.38.1
